mediaprocess: CB2 change default svc enable flag. [1/1]
PD#SWPL-144286

Problem:
svc function is default disable for 521.

Solution:
change default svc enable flag.

Verify:
t7c

+++ mediaprocess/include/AmlVencParamIntf.h
@@ -78,7 +78,7 @@ typedef struct _VencParam {
     int FrameRate;
     int PixFormat;
     bool SyncFrameRequest;
-    int SyncFramePeriod;
+    uint64_t SyncFramePeriod;
     stQpInfo QpInfo;
     int LayerCnt;
     bool prepend_header_flag;
@@ -111,7 +111,7 @@ public:
     virtual void GetVencParam(stVencParam &Param) = 0;
     virtual void SetSize(int width,int height) = 0;
     virtual void SetBitrate(int bitrate) = 0;
-    virtual void SetSyncPeriodSize(int Period) = 0;
+    virtual void SetSyncPeriodSize(uint64_t Period) = 0;
     virtual void SetFrameRate(int FrameRate) = 0;
     virtual void SetQpInfo(stQpInfo QpInfo) = 0;
     virtual void SetLayerCnt(int LayerCount) = 0;
